chore: bump pangolin-parachain spec
correct parachain direction
