+++ packages/apps/bridges/parachain-substrate/index.ts
@@ -1,2 +1,2 @@
-export * from './Parachain2Substrate';
 export * from './Substrate2Parachain';
+export * from './Parachain2Substrate';
