+++ packages/apps/bridges/register.tsx
@@ -5,7 +5,7 @@ import { parachainCrab, parachainPangolin } from 'shared/config/bridges/parachai
 import { crabCrabDVM, pangolinPangolinDVM } from 'shared/config/bridges/substrate-dvm';
 import { unknownUnavailable } from 'shared/config/bridges/unknown-unavailable';
 import { Darwinia2Ethereum, Ethereum2Darwinia } from './ethereum-darwinia';
-import { Parachain2Substrate, Substrate2Parachain } from './parachain-substrate';
+import { Substrate2Parachain, Parachain2Substrate } from './parachain-substrate';
 import { DVM2Substrate, Substrate2DVM } from './substrate-dvm';
 import { Substrate2SubstrateDVM, SubstrateDVM2Substrate } from './substrate-substrateDVM';
 import { Unavailable2Unknown, Unknown2Unavailable } from './unknown-unavailable';
@@ -47,7 +47,7 @@ pangolinPangolinDVM.setRedeemComponents(DVM2Substrate as FunctionComponent);
  * crab <-> crab parachain
  * pangolin <-> pangolin parachain
  */
-parachainCrab.setIssuingComponents(Parachain2Substrate as FunctionComponent);
-parachainCrab.setRedeemComponents(Substrate2Parachain as FunctionComponent);
-parachainPangolin.setIssuingComponents(Parachain2Substrate as FunctionComponent);
-parachainPangolin.setRedeemComponents(Substrate2Parachain as FunctionComponent);
+parachainCrab.setIssuingComponents(Substrate2Parachain as FunctionComponent);
+parachainCrab.setRedeemComponents(Parachain2Substrate as FunctionComponent);
+parachainPangolin.setIssuingComponents(Substrate2Parachain as FunctionComponent);
+parachainPangolin.setRedeemComponents(Parachain2Substrate as FunctionComponent);
